envision_product/chat/backend/requirements.txt
# requirements_enhanced.txt
# Core LangChain
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5

# Vector Store
chromadb>=0.4.0

# Document Processing
pypdf>=3.0.0
python-docx>=0.8.11
unstructured>=0.10.0
pandas>=1.5.0

# Additional Retrievers
rank-bm25>=0.2.2
sentence-transformers>=2.2.0

# Environment & Utils
python-dotenv>=1.0.0
numpy>=1.24.0
tqdm>=4.65.0

# OpenAI
openai>=1.0.0

# FastAPI & Web Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6  # For file uploads
pydantic>=2.5.0 # Explicitly define, though FastAPI pulls it

# Optional: For better text processing
nltk>=3.8
spacy>=3.6.0